token prob

+++ backend/foodgram/api/views.py
@@ -1,4 +1,5 @@
-from django.shortcuts import render, get_object_or_404
+from django.shortcuts import get_object_or_404
+from django.http import HttpResponse
 from rest_framework import permissions, viewsets
 from rest_framework.decorators import action
 from rest_framework.response import Response
@@ -10,6 +11,7 @@ from .permissions import IsAuthorOrReadOnly, IsAdminOrReadOnly
 from .serializers import (TagSerializer, IngredientSerializer,
                           RecipeSerializer, MiniRecipesSerializer,
                           ShoppingcartSerializer)
+from .services import shoppingcart
 
 
 class TagViewSet(viewsets.ModelViewSet):
@@ -29,46 +31,50 @@ class RecipeViewSet(viewsets.ModelViewSet):
     serializer_class = RecipeSerializer
     permission_classes = [IsAuthorOrReadOnly, IsAdminOrReadOnly]
 
-    @action(detail=True, methods=['post'],
+    @action(detail=True, methods=['post', 'delete'],
             permission_classes=[permissions.IsAuthenticated])
-    def recipe_fav(self, request, pk):
+    def favorite(self, request, pk):
         user = request.user
-        if Favorite.objects.filter(user=user, recipe=pk).exists():
-            return Response({"error": "уже в Избранном"},
+        if request.method == 'POST':
+            if Favorite.objects.filter(user=user, recipe=pk).exists():
+                return Response({"error": "уже в Избранном"},
+                                status=HTTP_400_BAD_REQUEST)
+            Favorite.objects.create(user=user, recipe=pk)
+            fav_recipe = get_object_or_404(Recipe, id=pk)
+            serializer = MiniRecipesSerializer(fav_recipe)
+            return Response(serializer.data, status=HTTP_201_CREATED)
+        elif request.method == 'DELETE':
+            if Favorite.objects.filter(user=user, recipe=pk).exists():
+                Favorite.objects.filter(user=user, recipe=pk).delete()
+                return Response(status=HTTP_200_OK)
+            return Response({"error": "not in Favorites"},
                             status=HTTP_400_BAD_REQUEST)
-        Favorite.objects.create(user=user, recipe=pk)
-        fav_recipe = get_object_or_404(Recipe, id=pk)
-        serializer = MiniRecipesSerializer(fav_recipe)
-        return Response(serializer.data, status=HTTP_201_CREATED)
 
-    @action(detail=True, methods=['delete'],
+    @action(detail=True, methods=['post', 'delete'],
             permission_classes=[permissions.IsAuthenticated])
-    def recipe_unfav(self, request, pk):
+    def shopping_cart(self, request, pk):
         user = request.user
-        if Favorite.objects.filter(user=user, recipe=pk).exists():
-            Favorite.objects.filter(user=user, recipe=pk).delete()
-            return Response(status=HTTP_200_OK)
-        return Response({"error": "not in Favorites"},
-                        status=HTTP_400_BAD_REQUEST)
-
-    @action(detail=True, methods=['post'],
-            permission_classes=[permissions.IsAuthenticated])
-    def shop_cart_add(self, request, pk):
-        user = request.user
-        if Shoppingcart.objects.filter(user=user, recipe=pk).exists():
-            return Response({"error": "уже в Корзине"},
+        if request.method == 'POST':
+            if Shoppingcart.objects.filter(user=user, recipe=pk).exists():
+                return Response({"error": "уже в Корзине"},
+                                status=HTTP_400_BAD_REQUEST)
+            Shoppingcart.objects.create(user=user, recipe=pk)
+            shop_cart = get_object_or_404(Recipe, id=pk)
+            serializer = ShoppingcartSerializer(shop_cart)
+            return Response(serializer.data, status=HTTP_201_CREATED)
+        elif request.method == 'DELETE':
+            if Shoppingcart.objects.filter(user=user, recipe=pk).exists():
+                Shoppingcart.objects.filter(user=user, recipe=pk).delete()
+                return Response(status=HTTP_200_OK)
+            return Response({"error": "not in Shopping Cart"},
                             status=HTTP_400_BAD_REQUEST)
-        Shoppingcart.objects.create(user=user, recipe=pk)
-        shop_cart = get_object_or_404(Recipe, id=pk)
-        serializer = ShoppingcartSerializer(shop_cart)
-        return Response(serializer.data, status=HTTP_201_CREATED)
- 
-    @action(detail=True, methods=['delete'],
+
+    @action(detail=False, methods=['get'],
             permission_classes=[permissions.IsAuthenticated])
-    def shop_cart_del(self, request, pk):
-        user = request.user
-        if Shoppingcart.objects.filter(user=user, recipe=pk).exists():
-            Shoppingcart.objects.filter(user=user, recipe=pk).delete()
-            return Response(status=HTTP_200_OK)
-        return Response({"error": "not in Shopping Cart"},
-                        status=HTTP_400_BAD_REQUEST)
+    def download_shopping_cart(self, request):
+        shoping_cart = shoppingcart(request)
+        response = HttpResponse(shoping_cart,
+                                content_type="text.txt; charset=utf-8")
+        filename = 'loaded_ingr.txt'
+        response["Content-Disposition"] = f"attachment; filename={filename}"
+        return response
